+++ src/Files.App/Files.App.csproj
@@ -104,7 +104,7 @@
         <ProjectReference Include="..\Files.Core\Files.Core.csproj" />
         <ProjectReference Include="..\Files.Sdk.Storage\Files.Sdk.Storage.csproj" />
         <ProjectReference Include="..\Files.Shared\Files.Shared.csproj" />
-        <ProjectReference Include="..\Files.SourceGenerator\Files.SourceGenerator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
+        <ProjectReference Include="..\Files.Core.SourceGenerator\Files.Core.SourceGenerator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
     </ItemGroup>
 
     <ItemGroup>
