+++ src/Files.Core.SourceGenerator/DependencyPropertyGenerator.cs
@@ -7,9 +7,9 @@ using System.Text;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
-using static Files.SourceGenerator.Utilities.SourceGeneratorHelper;
+using static Files.Core.SourceGenerator.Utilities.SourceGeneratorHelper;
 
-namespace Files.SourceGenerator
+namespace Files.Core.SourceGenerator
 {
 	[Generator]
 	public class DependencyPropertyGenerator : TypeWithAttributeGenerator
