+++ src/Files.Core.SourceGenerator/Files.Core.SourceGenerator.csproj
